refactor: unify shader logic and simplifies instanciation
