+++ source/Age/RenderPasses/CanvasBaseRenderGraphPass.cs
@@ -34,7 +34,7 @@ public abstract partial class CanvasBaseRenderGraphPass(VulkanRenderer renderer,
         {
             if (resource.Enabled)
             {
-                this.CommandBuffer.BindPipeline(resource.Pipeline);
+                this.CommandBuffer.BindShader(resource.Shader);
                 this.CommandBuffer.BindVertexBuffer(resource.VertexBuffer);
                 this.CommandBuffer.BindIndexBuffer(resource.IndexBuffer);
